[C] Fix warning.

+++ aeron-driver/src/test/c/aeron_packet_timestamps_test.cpp
@@ -298,7 +298,7 @@ TEST_F(PacketTimestampsTest, shouldPutTimestampInMessagesReservedValueWithMerged
         pubAPosition,
         pubAConstants.position_bits_to_shift,
         pubAConstants.initial_term_id);
-    int32_t termOffset = pubAPosition & (pubAConstants.term_buffer_length - 1);
+    int32_t termOffset = (int32_t)(pubAPosition & (pubAConstants.term_buffer_length - 1));
 
     std::stringstream publicationBStream;
     publicationBStream << destinationB;
